Graph algo refactor (#12)
* Split up knowledge graph ts
* Put graph algos in separate file
